// frontend/src/api/models/getTodoItemsStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Inspirate Consulting API
 * API for the Inspirate Consulting application
 * OpenAPI spec version: 1.0.0
 */

export type GetTodoItemsStatus = typeof GetTodoItemsStatus[keyof typeof GetTodoItemsStatus];


export const GetTodoItemsStatus = {
  all: 'all',
  completed: 'completed',
  incomplete: 'incomplete',
} as const;