// frontend/src/api/models/todoItemPage.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Inspirate Consulting API
 * API for the Inspirate Consulting application
 * OpenAPI spec version: 1.0.0
 */
import type { TodoItem } from './todoItem';

export interface TodoItemPage {
  /** A URL to the JSON Schema for this object. */
  readonly $schema?: string;
  /** @nullable */
  items: TodoItem[] | null;
  /**
     * Pass as cursor to get the next page; null when there are none
     * @nullable
     */
  next_cursor: string | null;
}